doubel

+++ lib/minus.h
@@ -1,6 +1,6 @@
 #include <bits/stdc++.h>
 using namespace std;
  
-int minus(int x,int y){
+double minus(int x,int y){
     return x - y;
 }
